Move badelf workflows to badelf app
Some workflows that only run badelf were located in the warren_lab app instead of the badelf app. The badelf model was also located in warrenapp for no particular reason. These are moved over in this commit.
