+++ src/simmate/apps/badelf/workflows/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .badelf import BadElf__Badelf__Badelf
